feat: devide component into presenter and container
