+++ package.json
@@ -36,7 +36,8 @@
     "test": "jest",
     "lint": "eslint src/**/*.js",
     "fmt": "prettier --write src/**/*.js",
-    "s": "yarn cache clean && react-native run-ios"
+    "s":
+      "yarn cache clean && REACT_DEBUGGER=\"unset ELECTRON_RUN_AS_NODE && open -g 'rndebugger://set-debugger-loc?port=19001' ||\" react-native run-ios"
   },
   "jest": {
     "preset": "jest-expo"
