+++ src/containers/PaymentAdd.js
@@ -1,7 +1,4 @@
-/**
- * @flow
- */
-// /* eslint-disable */
+// @flow
 
 import React, { Component } from 'react'
 import styled from 'styled-components'
